Refactor AuthActivity.kt. Crate Validator.kt, Constants.kt and Extensions.kt
