+++ .idea/misc.xml
@@ -16,7 +16,7 @@
         <entry key="app/src/main/res/drawable/linkedin.xml" value="0.2005" />
         <entry key="app/src/main/res/drawable/vk.xml" value="0.2005" />
         <entry key="app/src/main/res/font/open_sans_semibold.xml" value="0.2267361111111111" />
-        <entry key="app/src/main/res/layout/activity_auth.xml" value="0.165" />
+        <entry key="app/src/main/res/layout/activity_auth.xml" value="0.25863189697265626" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.25" />
         <entry key="app/src/main/res/layout/activity_profile.xml" value="0.165" />
         <entry key="app/src/main/res/layout/frame_profile.xml" value="0.25" />
