+++ app/src/main/java/com/example/level1/ProfileActivity.kt
@@ -3,6 +3,7 @@ package com.example.level1
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.level1.databinding.ActivityProfileBinding
+import com.example.level1.utils.Constants
 
 class ProfileActivity : AppCompatActivity() {
 
@@ -13,6 +14,6 @@ class ProfileActivity : AppCompatActivity() {
         binding = ActivityProfileBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
-        binding.myProfileName.text = intent.getStringExtra(AuthActivity.NAME_KEY)
+        binding.myProfileName.text = intent.getStringExtra(Constants.NAME_KEY)
     }
 }
